notification/requirements.txt
pyyaml>=6.0
nats-py>=2.7.0
python-telegram-bot>=22.0 